recoverseg/addmirrors/movemirrors: Add behave and unit tests
This commit adds/improves our code coverage for the following newly
added features:
1. Start the mirrors independently of each other
2. Improved console output in case any operation including
pg_basebackup/pg_rewind or start fails for one or more mirrors

+++ concourse/scripts/run_behave_test.sh
@@ -26,7 +26,11 @@ if gpstate > /dev/null 2>&1 ; then
 fi
 
 # TODO remove this pip install. only for debugging
-gpssh -f ~/segment_host_list -e 'pip3 install coverage --user'
+if [ -f ~/segment_host_list ]; then
+    gpssh -f ~/segment_host_list -e 'pip3 install coverage --user'
+fi
+
+ssh sdw5 'pip3 install coverage --user' > /dev/null 2>&1 || true
 
 cd /home/gpadmin/gpdb_src/gpMgmt
 make -f Makefile.behave behave flags="$BEHAVE_FLAGS"
